refactor server code
