+++ assets/exchange-rates/latest.json
@@ -1,176 +1,176 @@
 {
   "disclaimer": "Usage subject to terms: https://openexchangerates.org/terms",
   "license": "https://openexchangerates.org/license",
-  "timestamp": 1508929200,
+  "timestamp": 1509472800,
   "base": "USD",
   "rates": {
-    "AED": 3.672952,
-    "AFN": 68.497,
-    "ALL": 113.3,
-    "AMD": 482,
-    "ANG": 1.78506,
+    "AED": 3.6728,
+    "AFN": 68.30326,
+    "ALL": 114.62,
+    "AMD": 482.71,
+    "ANG": 1.785454,
     "AOA": 165.9225,
-    "ARS": 17.5025,
-    "AUD": 1.296871,
-    "AWG": 1.789995,
+    "ARS": 17.655,
+    "AUD": 1.305154,
+    "AWG": 1.790501,
     "AZN": 1.6985,
-    "BAM": 1.66205,
+    "BAM": 1.6789,
     "BBD": 2,
-    "BDT": 83.101,
-    "BGN": 1.661511,
-    "BHD": 0.377122,
-    "BIF": 1750.8,
+    "BDT": 83.044,
+    "BGN": 1.67843,
+    "BHD": 0.377164,
+    "BIF": 1755,
     "BMD": 1,
-    "BND": 1.363256,
-    "BOB": 6.910039,
-    "BRL": 3.242187,
+    "BND": 1.3624,
+    "BOB": 6.910011,
+    "BRL": 3.279886,
     "BSD": 1,
-    "BTC": 0.000179480339,
-    "BTN": 64.97899,
-    "BWP": 10.389893,
-    "BYN": 1.95198,
-    "BZD": 2.00999,
-    "CAD": 1.271623,
-    "CDF": 1562.881563,
-    "CHF": 0.992777,
-    "CLF": 0.02349,
-    "CLP": 632.4,
-    "CNH": 6.646915,
-    "CNY": 6.6446,
-    "COP": 2961.52,
-    "CRC": 569.835,
+    "BTC": 0.000157092355,
+    "BTN": 64.831491,
+    "BWP": 10.534446,
+    "BYN": 1.977088,
+    "BZD": 2.010437,
+    "CAD": 1.289762,
+    "CDF": 1576,
+    "CHF": 0.996566,
+    "CLF": 0.02363,
+    "CLP": 635.9,
+    "CNH": 6.630907,
+    "CNY": 6.6382,
+    "COP": 3042.44,
+    "CRC": 569.97,
     "CUC": 1,
     "CUP": 25.5,
-    "CVE": 94,
-    "CZK": 21.7166,
+    "CVE": 95.145,
+    "CZK": 22.031549,
     "DJF": 178.97,
-    "DKK": 6.320475,
-    "DOP": 48.026173,
-    "DZD": 114.4725,
-    "EGP": 17.6483,
+    "DKK": 6.389039,
+    "DOP": 48.025,
+    "DZD": 115.096723,
+    "EGP": 17.6503,
     "ERN": 15.191062,
-    "ETB": 27.256793,
-    "EUR": 0.849135,
-    "FJD": 2.065344,
-    "FKP": 0.75514,
-    "GBP": 0.75514,
-    "GEL": 2.501054,
-    "GGP": 0.75514,
-    "GHS": 4.381,
-    "GIP": 0.75514,
-    "GMD": 47.375,
-    "GNF": 9023.7,
-    "GTQ": 7.352554,
-    "GYD": 207.578977,
-    "HKD": 7.8049,
-    "HNL": 23.608851,
-    "HRK": 6.379215,
-    "HTG": 63.162196,
-    "HUF": 263.3834,
-    "IDR": 13551.246714,
-    "ILS": 3.509748,
-    "IMP": 0.75514,
-    "INR": 64.895,
-    "IQD": 1167,
-    "IRR": 34305.446723,
-    "ISK": 105.82515,
-    "JEP": 0.75514,
-    "JMD": 127.303036,
+    "ETB": 27.325,
+    "EUR": 0.858552,
+    "FJD": 2.079006,
+    "FKP": 0.753377,
+    "GBP": 0.753377,
+    "GEL": 2.538029,
+    "GGP": 0.753377,
+    "GHS": 4.4,
+    "GIP": 0.753377,
+    "GMD": 47.45,
+    "GNF": 9000,
+    "GTQ": 7.354253,
+    "GYD": 208.035254,
+    "HKD": 7.80155,
+    "HNL": 23.520692,
+    "HRK": 6.459524,
+    "HTG": 63.164696,
+    "HUF": 267.2765,
+    "IDR": 13541.246714,
+    "ILS": 3.52047,
+    "IMP": 0.753377,
+    "INR": 64.6565,
+    "IQD": 1183,
+    "IRR": 34571.446723,
+    "ISK": 105.48,
+    "JEP": 0.753377,
+    "JMD": 127.147067,
     "JOD": 0.709001,
-    "JPY": 114.1445,
-    "KES": 103.9,
-    "KGS": 68.516981,
-    "KHR": 4043.35,
-    "KMF": 419.194527,
+    "JPY": 113.68357143,
+    "KES": 103.8,
+    "KGS": 68.62998,
+    "KHR": 4039,
+    "KMF": 423.125,
     "KPW": 900,
-    "KRW": 1124.9925,
-    "KWD": 0.302346,
-    "KYD": 0.833317,
-    "KZT": 333.64,
-    "LAK": 8302.1,
-    "LBP": 1515.749713,
-    "LKR": 153.84,
-    "LRD": 118.75,
-    "LSL": 13.722298,
-    "LYD": 1.365489,
-    "MAD": 9.424704,
-    "MDL": 17.296644,
-    "MGA": 3088.7,
-    "MKD": 52.275,
-    "MMK": 1354,
-    "MNT": 2453.31916,
-    "MOP": 8.038668,
-    "MRO": 355.17,
-    "MUR": 34.0535,
-    "MVR": 15.400126,
-    "MWK": 726.17,
-    "MXN": 19.223922,
-    "MYR": 4.235674,
-    "MZN": 61.195819,
-    "NAD": 13.7525,
-    "NGN": 360.5,
-    "NIO": 30.745369,
-    "NOK": 8.011145,
-    "NPR": 104.076159,
-    "NZD": 1.453225,
-    "OMR": 0.38499,
+    "KRW": 1118.24,
+    "KWD": 0.30262,
+    "KYD": 0.83346,
+    "KZT": 334.96,
+    "LAK": 8320,
+    "LBP": 1510,
+    "LKR": 153.671111,
+    "LRD": 119.375,
+    "LSL": 14.145,
+    "LYD": 1.36586,
+    "MAD": 9.498694,
+    "MDL": 17.329616,
+    "MGA": 3160,
+    "MKD": 52.845,
+    "MMK": 1365.66006,
+    "MNT": 2451.989147,
+    "MOP": 8.036067,
+    "MRO": 355,
+    "MUR": 34.2285,
+    "MVR": 15.409873,
+    "MWK": 725.5,
+    "MXN": 19.131385,
+    "MYR": 4.233523,
+    "MZN": 61,
+    "NAD": 14.14125,
+    "NGN": 360,
+    "NIO": 30.55,
+    "NOK": 8.174841,
+    "NPR": 103.61,
+    "NZD": 1.461691,
+    "OMR": 0.384262,
     "PAB": 1,
-    "PEN": 3.236504,
-    "PGK": 3.202258,
-    "PHP": 51.879,
-    "PKR": 105.289978,
-    "PLN": 3.595958,
-    "PYG": 5630.55,
-    "QAR": 3.79235,
-    "RON": 3.905236,
-    "RSD": 101.548179,
-    "RUB": 57.5429,
-    "RWF": 853.695,
-    "SAR": 3.750534,
-    "SBD": 7.800511,
-    "SCR": 13.769,
-    "SDG": 6.676877,
-    "SEK": 8.226241,
-    "SGD": 1.362583,
-    "SHP": 0.75514,
+    "PEN": 3.2477,
+    "PGK": 3.21265,
+    "PHP": 51.605,
+    "PKR": 105.335,
+    "PLN": 3.641189,
+    "PYG": 5648.35,
+    "QAR": 3.79998,
+    "RON": 3.952716,
+    "RSD": 102.33,
+    "RUB": 58.3229,
+    "RWF": 841,
+    "SAR": 3.7502,
+    "SBD": 7.792911,
+    "SCR": 13.903667,
+    "SDG": 6.675,
+    "SEK": 8.373662,
+    "SGD": 1.362512,
+    "SHP": 0.753377,
     "SLL": 7643.328121,
-    "SOS": 578.505,
+    "SOS": 593,
     "SRD": 7.448,
     "SSP": 130.2634,
-    "STD": 20799.099895,
-    "SVC": 8.749776,
+    "STD": 20915.200481,
+    "SVC": 8.751478,
     "SYP": 514.98999,
-    "SZL": 13.726064,
-    "THB": 33.198333,
-    "TJS": 8.805036,
-    "TMT": 3.509961,
-    "TND": 2.469012,
-    "TOP": 2.224732,
-    "TRY": 3.731964,
-    "TTD": 6.730093,
-    "TWD": 30.25249,
-    "TZS": 2246,
-    "UAH": 26.69024,
-    "UGX": 3665.05,
+    "SZL": 14.145,
+    "THB": 33.22,
+    "TJS": 8.812212,
+    "TMT": 3.499986,
+    "TND": 2.495606,
+    "TOP": 2.269382,
+    "TRY": 3.794075,
+    "TTD": 6.731608,
+    "TWD": 30.138,
+    "TZS": 2248.5,
+    "UAH": 26.86611,
+    "UGX": 3655.9,
     "USD": 1,
-    "UYU": 29.65572,
-    "UZS": 8068.7,
-    "VEF": 10.07525,
-    "VND": 22722.019974,
-    "VUV": 105.966667,
-    "WST": 2.515073,
-    "XAF": 556.996139,
-    "XAG": 0.05919278,
-    "XAU": 0.00078533,
+    "UYU": 29.244046,
+    "UZS": 8090,
+    "VEF": 10.59115,
+    "VND": 22711.553308,
+    "VUV": 106.763333,
+    "WST": 2.523173,
+    "XAF": 563.173487,
+    "XAG": 0.05985882,
+    "XAU": 0.00078725,
     "XCD": 2.70255,
-    "XDR": 0.70996,
-    "XOF": 556.996139,
-    "XPD": 0.00104059,
-    "XPF": 101.328776,
-    "XPT": 0.00109078,
-    "YER": 250.306642,
-    "ZAR": 13.751559,
-    "ZMW": 9.895,
+    "XDR": 0.711903,
+    "XOF": 563.173487,
+    "XPD": 0.00101835,
+    "XPF": 102.452559,
+    "XPT": 0.00108815,
+    "YER": 250.275,
+    "ZAR": 14.137012,
+    "ZMW": 10.015,
     "ZWL": 322.355011
   }
 }
