+++ server/locales.js
@@ -1,0 +1,7 @@
+var locales = {};
+
+Object.keys(require('../public/locales.json')).forEach(function(locale) {
+  locales[locale.toLowerCase()] = locale;
+});
+
+module.exports = locales;
